Update MysqlDriver.php
The installation error log told me that the current version of MariaDB that I'm using wasn't matching with the minimum version, however, it wasn't comparing it to a version that I'm using but instead it was comparing the minimum version to a string of "MariaDB". On line 74, I switched the variables around in the array brackets when the explode function returned the list of strings. Once the change was made, the installation went successfully.